third_party/WebKit/LayoutTests/fast/viewport/viewport-2-expected.txt

ALERT: viewport size 980x480 scale 0.326531 with limits [0.326531, 5] and userScalable true
